pyproject.toml
[tool.poetry]
name = "uhoo-homeassistant"
version = "0.0.7"
description = "uHoo integration for Home Assistant"
authors = ["Christopher Sacca <csacca@csacca.net>"]
repository = "https://github.com/csacca/uhoo-homeassistant"
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
pyuhoo = {git = "https://github.com/wrouesnel/pyuhoo.git"}

[tool.poetry.group.dev.dependencies]
setuptools = "*"

pytest = "*"
pytest-homeassistant-custom-component = "^0.13.107"

flake8 = "*"
flake8-docstrings = "*"
pydocstyle = "*"
black = "*"
isort = "*"
mypy = "*"

[tool.isort]
profile = "black"
indent = "    "
force_sort_within_sections = true
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
default_section = "THIRDPARTY"
known_first_party = ["custom_components", "tests" , "homeassistant"]
forced_separate = "tests"
combine_as_imports = true

[tool.mypy]
pretty = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--cov custom_components.uhoo"
console_output_style = "count"
timeout = "9"
testpaths = ["tests"]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
